requirements.txt
streamlit
python-dotenv
requests
pydantic
numpy
pandas

langchain
langchain-community
langchain-core

faiss-cpu
sentence-transformers

sympy

pillow
pytesseract
easyocr

openai-whisper

sounddevice
scipy

pyyaml
tqdm
rich
loguru